Added infrastructure code for processing commands from clients.
Added code to handle commands from clients in leader. A logEntry is
created for the command and is appended to the leader log. The log gets
propagated to other nodes in the cluster during its next heartbeat. Each
node upon receipt of the AppendEntry request will update its local log.