Fix client command processing. Execute command when corresponding log is committed and respond to the client.
On client command invocation, processCommand in prez core makes initial
validation and invokes clusterProcessCommand, where a logEntry is
created. Note that we simply fake error in processCommand so that the
client doesn't get reset. Now, once majority cluster nodes have appended
the entry, when the log is committed, the command handler is invoked. A
hash table(log index,prezClient pointer) is introduced to retrieve the
prezClient later to add the reply to the client. This is applicable only
for leader and the followers should simply execute the command (pending implementation).